refactor: enum package 위치 변경
