+++ src/main/java/com/ll/hype/domain/customer/customer/entity/CustomerQ.java
@@ -1,6 +1,5 @@
 package com.ll.hype.domain.customer.customer.entity;
 
-import com.ll.hype.global.enums.QuestionCategory;
 import com.ll.hype.global.jpa.BaseEntity;
 import com.ll.hype.domain.member.member.entity.Member;
 import jakarta.persistence.Entity;
