+++ src/main/java/com/ll/hype/domain/customer/customer/entity/QuestionCategory.java
@@ -1,4 +1,4 @@
-package com.ll.hype.global.enums;
+package com.ll.hype.domain.customer.customer.entity;
 
 public enum QuestionCategory {
